requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pyhd3eb1b0_0
blas=1.0=openblas
bottleneck=1.3.5=py311hbed6279_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py311h6a678d5_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2023.11.17=py311h06a4308_0
cffi=1.16.0=py311h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py311h06a4308_0
contourpy=1.2.0=py311hdb19cb5_0
cryptography=41.0.3=py311hdda0065_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
eigen=3.3.7=hd09550d_1
expat=2.5.0=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.13.1=py311h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.9.2=py311h06a4308_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.37=py311h06a4308_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
harfbuzz=4.3.0=hf52aaf7_1
hdf5=1.12.1=h2b7332f_3
icu=58.2=he6710b0_3
idna=3.4=py311h06a4308_0
jinja2=3.1.2=py311h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py311h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libcurl=8.4.0=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libopenblas=0.3.21=h043d6bf_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py311h5eee18b_0
matplotlib=3.8.0=py311h06a4308_0
matplotlib-base=3.8.0=py311ha02d727_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py311h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.7=py311h812550d_0
numpy=1.26.0=py311h24aa872_0
numpy-base=1.26.0=py311hbfb1bba_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv=4.6.0=py311h10ae9b0_5
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
packaging=23.1=py311h06a4308_0
pandas=2.1.1=py311ha02d727_0
pathtools=0.1.2=pyhd3eb1b0_1
pcre=8.45=h295c915_0
pillow=10.0.1=py311ha6cbd5a_0
pip=23.3=py311h06a4308_0
pixman=0.40.0=h7f8727e_1
ply=3.11=py311h06a4308_0
protobuf=3.20.3=py311h6a678d5_0
psutil=5.9.0=py311h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py311h06a4308_0
pyparsing=3.0.9=py311h06a4308_0
pyqt=5.15.10=py311h6a678d5_0
pyqt5-sip=12.13.0=py311h5eee18b_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.5=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=py311h06a4308_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py311h06a4308_0
pyyaml=6.0.1=py311h5eee18b_0
qt-main=5.15.2=h7358343_9
qt-webengine=5.15.9=h9ab4d14_7
qtwebkit=5.212=h3fafdc1_5
readline=8.2=h5eee18b_0
requests=2.31.0=py311h06a4308_0
sentry-sdk=1.9.0=py311h06a4308_0
setproctitle=1.2.2=py311h5eee18b_0
setuptools=68.0.0=py311h06a4308_0
sip=6.7.12=py311h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py311h06a4308_0
tk=8.6.12=h1ccaba5_0
torch=2.1.1=pypi_0
torchvision=0.16.1=pypi_0
tornado=6.3.3=py311h5eee18b_0
tqdm=4.65.0=py311h92b7b1e_0
triton=2.1.0=pypi_0
typing-extensions=4.7.1=py311h06a4308_0
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.18=py311h06a4308_0
wandb=0.15.12=pyhd8ed1ab_0
wheel=0.41.2=py311h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
